fix : 정확한 실수 표현을 위한 BigDecimal로 타입 변경

+++ src/main/java/calculator/calculator/Calculator.java
@@ -4,15 +4,17 @@ import calculator.calculator.formula.FormulaParser;
 import calculator.calculator.notation.calculation.NotationCalculation;
 import calculator.calculator.notation.parser.NotationParser;
 
+import java.math.BigDecimal;
+
 public abstract class Calculator {
 
     NotationCalculation calculation;
     NotationParser notationParser;
     FormulaParser formulaParser;
 
-    public Double calculate(String formula) {
+    public BigDecimal calculate(String formula) {
         return calculation.calculate(
-                notationParser.parseFrom(
-                        formulaParser.parseFrom(formula)));
+                        notationParser.parseFrom(
+                                formulaParser.parseFrom(formula)));
     }
 }
