+++ src/main/java/calculator/calculator/notation/calculation/NotationCalculation.java
@@ -1,7 +1,8 @@
 package calculator.calculator.notation.calculation;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface NotationCalculation {
-    Double calculate(List<String> formulas);
+    BigDecimal calculate(List<String> formulas);
 }
